+++ src/http/response/defs.rs
@@ -38,9 +38,9 @@ pub trait MappableResponse<T>: Sized {
 }
 
 pub trait StructResponse: Sized {
-    fn send_response(self) -> HttpResponse;
+    fn into_response(self, suc: &str) -> HttpResponse;
 
-    fn send_struct_result(self) -> Result<HttpResponse, AppMessage>;
+    fn respond(self, suc: &str) -> Result<HttpResponse, AppMessage>;
 }
 
 pub trait OptionResultResponse<T> {